Make sure we can build ndt-server on macOS
We don't support macOS officially. However it's useful to build on
a system that is different from the containers where we deploy.

As mentioned in #100, I am convinced that Travis is different enough
from the containers where we deploy to look like macOS.
