+++ bbr/bbr_stub.go
@@ -4,12 +4,14 @@ package bbr
 
 import (
 	"os"
+
+	"github.com/m-lab/ndt-server/ndt7/model"
 )
 
 func enableBBR(*os.File) error {
 	return ErrNoSupport
 }
 
-func getMaxBandwidthAndMinRTT(*os.File) (float64, float64, error) {
-	return 0.0, 0.0, ErrNoSupport
+func getMaxBandwidthAndMinRTT(*os.File) (model.BBRInfo, error) {
+	return model.BBRInfo{}, ErrNoSupport
 }
